Move photoscan file out of db folder (#180)
