+++ src/openlifu/db/__init__.py
@@ -1,11 +1,9 @@
 from .database import Database
-from .photoscan import Photoscan
 from .session import Session
 from .subject import Subject
 
 __all__ = [
     "Subject",
     "Session",
-    "Database",
-    "Photoscan"
+    "Database"
 ]
